// apps/app/shared/route-paths-type.ts
export enum RoutePathsType {
  Default = '/',
  Landing = '/landing',
  Home = '/home',
  Explore = '/explore',
  Signin = '/signin',
  Signup = '/signup',
  EmailConfirmation = '/email-confirmation',
  ForgotPassword = '/forgot-password',
  TermsOfService = '/terms-of-service',
  PrivacyPolicy = '/privacy-policy',
  Search = '/search',
  CompleteProfile = '/complete-profile',
  Customer = '/customer',
  Recents = '/recents',
  Player = '/player',
  Category = '/category',
  Collection = '/collection/:id',
  Affirmation = '/affirmation',
  Permissions = '/permissions',
  Help = '/help',
  ResetPassword = '/reset-password',
  Store = '/store',
  StoreWithId = '/store/:id',
  Events = '/events',
  Notifications = '/notifications',
  Cart = '/cart',
  Chats = '/chats',
  ChatsWithId = '/chats/:id',
  Checkout = '/checkout',
  Account = '/account',
  AccountWithId = '/account/:id',
  AccountWithIdLikes = '/account/:id/likes',
  AccountStatus = '/account/status',
  AccountStatusWithId = '/account/status/:id',
  AccountStatusWithIdFollowers = '/account/status/:id/followers',
  AccountStatusWithIdFollowing = '/account/status/:id/following',
  AccountEvents = '/account/events',
  AccountOrderHistory = '/account/order-history',
  AccountAddresses = '/account/addresses',
  AccountLikes = '/account/likes',
  AccountHelp = '/account/help',
  AccountAddFriends = '/account/add-friends',
  Settings = '/settings',
  SettingsAccount = '/settings/account',
  SettingsProfessionalAccount = '/settings/professional-account',
  OrderConfirmed = '/order/confirmed',
  OrderConfirmedWithId = '/order/confirmed/:id',
  ServerError = '/server-error',
}